backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AWS Lambda support (for Lambda deployment)
mangum==0.17.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Authentication & Security
pyjwt[crypto]==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pycognito==2023.5.0

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
shortuuid==1.0.11
email-validator==2.1.0